requirements/development.txt
-r base.txt
django-debug-toolbar>=1.4
Werkzeug>=0.11.2
django-allauth
django-basis
django-bootstrap3
django-extensions
django-guardian
django-pure-pagination
django-select-multiple-field
django-spurl
django-adv-cache-tag<1.0
django_filter
pytz